Switch namespace of directory services cache
The directory services cache pre-dates when we had a dedicated
generic plugin for directory services. Move it to the proper
namespace and clean up.
