+++ src/middlewared/middlewared/plugins/service_/services/pseudo/misc.py
@@ -201,11 +201,11 @@ class DSCacheService(PseudoServiceBase):
     name = "dscache"
 
     async def start(self):
-        await self.middleware.call('dscache.refresh')
+        await self.middleware.call('directoryservices.cache.refresh')
 
     async def stop(self):
         await self.middleware.call('idmap.clear_idmap_cache')
-        await self.middleware.call('dscache.refresh')
+        await self.middleware.call('directoryservices.cache.refresh')
 
 
 class UserService(PseudoServiceBase):
